Ensuring toga_core resource are included in distribution and correcting determination of toga_path
When attempting to use toga-demo on Ubuntu (Gnome) 16.10, I get an error on start up that is cannot load the Tiberius icon.

On investigation, it seemed that the toga_core resources are not included in the distribution package even though they are mentioned in `package_data` in the setp.py of `src/core`.

Doing some digging, it seems that `package_data` keys must be relative to a package `__init__.py`, which the `resources` folder is not (or at least not relative to the `toga` pacakge specified.

To resolve the issue I needed to move the `resources` folder into the `toga` package in `src/core`.

This also left the `toga_dir` in `src/core/toga/interface/icon.py` incorrect, as it assumes that the resources folder is under `/resources/` when it is now under `/toga/resources/`. So I corrected the path generation as well.

This appears to work correctly on Linux but I do not have other platforms available to confirm against.