Write tests for Font
The test file, test_font.py, now has 100% coverage over
`src/core/toga/font.py`.

A change was made to `src/gtk/toga_gtk/font.py` as well becuase its Font
class was attempting to access `self.family` and `self.size`, which
fails because those are properties of `self.interface` (which is an
instance of the the Toga Core API Font class, the one the tests were
written for).