Implement refresh() for GTK toga.Table
With this commit, row data given to toga.Table's constructor will be rendered,
and mutation to an instances's `data` property by `insert()`, `remove()`, and
`clear` will cause the native view to refresh.

A change is necessary to the Core API Table to cause a `refresh()` each time
the `data` property setter is called (such as during the constructor, if data
was given). Otherwise, the data is not shown in the view until one of the above
mentioned mutations is called.

A bug is fixed in the `create` method that caused each column to display the
same text.

Signed-off-by: Drew Meier <darius.montez@gmail.com>