Handle case where selected row is None + Revert incorrect docstring
In addition to the above changes, the Gtk.TreeSelection is used to track
changes to the selection, which simplifies the code since a reference to the
selection object is retained.

Another case is handled. When the on_select handler is replaced, the old one
is disconnected, and (now) the connection id is also removed from the list.
Previously, the signal was disconnected, but the connection id remained in the
list. When the next on_select was handled, an error was thrown since it was
trying to disconnect a signal that had already been disconnected.

Signed-off-by: Drew Meier <darius.montez@gmail.com>