Cleanup event handling for GTK+ Table and Button
It makes little sense to store a list of signal connection IDs if only one
handler can be attached at a time.

With this commit, a signal/event is connected to a private method, which in
turn calls the handler set by the interface, if it exists. To me, this feels
more natural (and even more readable).

Signed-off-by: Drew Meier <darius.montez@gmail.com>