gtk/ProgressBar: scrap GObject for custom interval function
The GTK ProgressBar example uses GObject.timeout_add to repeatedly
execute `.pulse()`. When the callback handler returns False, the
timeout is stopped; however, the progress bar seems for animate
slowly to a stop. This looks great and all, but it means that if
`set_fraction()` is called, it will be overriden by the next tick of
the timeout.

I wrote my own timeout/interval function using threading.Timer. The program
waits one interval duration after the interval is cancelled before calling
`set_fraction`, and the desired behavior is reached!

Signed-off-by: Drew Meier <darius.montez@gmail.com>