add comment to warn interested windows users
In my opinion toga should be more open about the state of the implementation.
Hidden explanations in issues do not help to get more native Windows users on board.
That might sound a bit harsh, but it was my experience. Strange logo, no demo  - only a fun project of some kids.

I now know that this not the case. But it will really need some help and convincing. Please...