Make Canvas handle explicit redraws in GTK
The Canvas widget has a `redraw` method that is meant to allow for
asking it to redraw itself programatically. The backend implementation
for it in GTK was so far missing. In practice this manifested itself in
having the results of method calls on the Canvas Widget not show up on
screen until a redraw event was triggered E.g. by resizing the window.

This PR adds implementation for the `redraw` method in GTK.

Signed-off-by: Barak Korren <bkorren@redhat.com>