Add a `resize` event to Canvas and a GTK impl.
When a windows containing a Canvas widget is resized, its usually
desirable to update the Canvas contents to match the new window size.
This patch provides a way to do so by introducing a `resize` event that
can be hooked into a callback.

The patch also includes an implementation of that event in GTK. It is
somewhat less straight-forward due to the fact that GTK does not seem to
include a `resize` event of its own.

Signed-off-by: Barak Korren <bkorren@redhat.com>