Added a more comprehensive Canvas example
The current Canvas example app is quite basic.

Adding a new example that makes an effort to utilize a wide range of
canvas capabilities. This app is meant to fulfill two purposes:
1. Help Toga users figure out how to use the more complex Canvas
   functions.
2. Allow comparing and testing the Canvas behaviour acroos different
   platforms.

Still keeping the existing example because given the complexity of this
one, its beneficial to a also have a very simple one.

Signed-off-by: Barak Korren <bkorren@redhat.com>