Add base classes for `Evented` objects
`Evented` objects are objects that expose certain events for which
callbacks can be set. Widgets that can have callbacks that are called
when users interact with them are a common example of this concept.

This commit only added the relevant base classes and tests thereof.
Future commits are expected to put then into use throughout the Toga
code base.

Discussion as to why this kind of change to the Toga code base may be
desirable can be found in issue #797.

Signed-off-by: Barak Korren <bkorren@redhat.com>