Fix dialogs app
This app does not reuse buttons in secondary window to avoid the bug related to winforms not having `remove_child` method (when reusing a button from the main window in the secondary window, the app is trying to remove it from the main window first). Also, the question dialog handling app exit is replaced by confirm dialog.

Signed-off-by: Olga <obulat@gmail.com>