Make changing Winforms imageview image possible
Currently, when we set an image to Winforms ` imageview`, it doesn't check if it already has the image. This way, when changing the image, resources are not being freed (disposed of), and we cannot delete the image that was previously used because of the Permission Error.
This PR adds a check for the existing image. If it exists, we free the resources by calling `Dispose`, and only then set the new image.