Align tutorial installation instructions with readme
The readme distinguishes between Ubuntu versions. Further the list of
debian packages in the readme differs from the one in the tutorial.
Because the readme is more recent (see 82c94a72) it's presumably more up
to date.

Hopefully this change doesn't drop a debian package which is used only
by the tutorial but not by the readme.