Update dependencies of new Ubuntu/Debian versions
The tests were performed with minimal docker containers for Ubuntu 18.04
and 20.04. The resources and instructions are available at
https://github.com/rknuus/toga-dependency-verification.

All 4 tutorial examples were tested by starting them and verifying that
the application window reacts to mouse/keyboard input. The most
sensitive example is the browser of tutorial 3 because (unlike the other
examples) it depends on WebKit2.