[Gtk] fix replacing a window content by another one
Currently, it doesn't work to set a content on a window if there is
already one, Gtk displays this error:

"Attempting to add a widget with type GtkBox to a GtkApplicationWindow,
but as a GtkBin subclass a GtkApplicationWindow can only contain one
widget at a time; it already contains a widget of type GtkBox"

This is because there is an existing Box layout inside the window.

To fix it, we remove the existing layout if it already exists.

Note that to work, it is required to call again 'window.show()' after
replacing the content.