Hide webview black background when resizing
By default in Windows, a webview will render a black background behind it. When resizing the webview, this black background becomes visible for a couple of frames while the html rendered on top is still in the process of updating, resulting in black "bars" appearing at the side of the webview when resizing it.

This can look quite ugly, see: https://youtu.be/Xj1rW4L-Pqk :) Hence, with this PR Im proposing to make it transparent.