Fix and tests for tree node modifications prior to expansion
Appending/inserting/changing/removing a tree data node prior to its expansion was throwing an AttributeError.  As an optimisation, objects representing tree nodes are created on first render. Trying to access these prior to tree expansion is causing the issue.

Added a node_impl() method that can be used anywhere that the user requests node._impl, and ensures that the impl exists.