Add NamedTuples for container Position and Size
Previously, Toga made extensive use of tuples for passing size and
position data.  The passing of this data assumes the user has certain
understanding, such as a position tuples having width as the first
argument rather than height.

Here, we update the Toga core to add NamedTuples to the core API as
values returned when accessing sizes and positions. Methods that accept
sizes or positions tuples as input will now accept both tuples and
NamedTuples to ensure backwards compability.