cocoa: fix default icon search when standalone app is not an .app bundle
When trying to obtain the default application icon in macOS standalone
application by querying `CFBundleIconFile` key, account for possibility
that application is not built as an .app bundle. I.e., ensure that the
returned string is not `None` before trying to construct `pathlib.Path`
with it. If returned string is `None`, raise `FileNotFoundError`, which
is handled by the calling layer.

This fixes `TypeError: argument should be a str or an os.PathLike object
where __fspath__ returns a str, not 'NoneType'` error when application
has no icon set and PyInstaller is used to build a POSIX executable
(as opposed to an .app bundle).