Ignore GNOME location settings
This matches Geoclue's own expectations regarding permissions:
https://gitlab.freedesktop.org/geoclue/geoclue/-/issues/111

TL;DR: Location permissions essentially do not meaningfully exist
outside of sandboxed applications on Linux.

Furthermore, attempting to respect GNOME's settings only creates
confusion for users, especially if GSettings is available on the
system but unused by the application user's DE (e.g., KDE). In
that case, the setting will be disabled (as that is the default),
and the user will have no insight into where or why the location
permissions are being denied to the application.