feat: Windows: Update canvas.py to use line_height_factor (#13)
Added a line_height parameter to write_text, _text_path and measure_text, in order for the user to be able to customize the line height of their text. Default value is currently 1.2 if no other value is specified.
This addition was done for the Windows version of the application.

Fixes #3