Rework lazy loading in toga core (#3359)
Rework how lazy loading works in toga core so that type checking tools have full
access to type information.

This works by doing the full import in a `.pyi` type annotation file, and then
using that file as a source of truth that can be *read*, but not imported or
parsed, to provide the map of classes that can be imported at runtime.