Make toga-winforms platform-specific
Exclude x86 and arm64 dlls from the wheel,
because they would not work anyway without pythonnet.
We decided to keep the dlls in the source.
MANIFEST.in controls what goes into the wheel.

Make use of bdist_wheel plat-name config to set the platform-tag to be "win_amd64" instead of "any".

Update toga-winforms keywords and classifiers to reflect reality.