Fix proactor event loop performance issue on Winforms.
Fixes the proactor event loop performance issue on Winforms by
utilizing an adaptive approach to determine when to poll the next
event loop iteration by checking for ready and/or scheduled events.

This change is tested using the demo provided at #2613, in addition
to running toga-demo and clicking the buttons.  The demo provided at
snippet of code beneath it; both run in about 13 seconds.

Fixes #2613.