Add API for current backend's name.
This adds a new function get_backend_and_factory that identifies the best
backend and loads its factory module, returning the backend name and the
factory.  The existing get_current_factory() function is refactored to be
a wrapper around the new function that simply returns just the factory.

In addition, lazy attributes 'backend' and 'factory' are added to the
toga and toga.package interfaces.

Fixes #3923.