Ensure Canvas Tests Explicitly Begin Paths
Currently most Canvas tests call begin_path befor each shape they draw,
or use context managers like Fill and Stroke which have equivalent behavior.
However two tests do not do that: test_transforms and test_ellipses.

This change brings those two tests into alignment with the other tests.

This is related to the discussion in #3993. Depending on how that is decided
all the begin_paths may not be needed, but for now it is better to be
consistent.