Fix overridden color handling in Qt backend
The Qt backend used to have the native widget autofill some widgets in create(),
but those widgets must manually be set to a default transparent background.  However,
this was broken because create() ran earlier than setting the default background colors.

This PR fixes this mechanism by providing 2 customization functions, one to override
color settings and one to override color roles, called at appropriate points in the
__init__ sequence to ensure that the override works.  The color role override is
not necessary at this stage but will be helpful when landing Selection.  Moved
all default color / color customization changes to the color settings override
function, as that'll run after and will not be overwritten by the
color role settings in __init__.