Fix NumberInput test on Cocoa
``test_increment_decrement`` of NumberInput does not work properly
on macOS Tahoe (with a Python compiled on previous macOS Versions;
I have not tested a Tahoe-compiled Python) because the increment probe
function sends a mouse-down event but never sends a mouse-up one,
and it is interpreted as holding the mouse to indefinitely increment
the NumberInput, and the handled signal never fires.  This PR adds
a mechanism to post an event without waiting for it to be handled
in order to post the mouse-up and mouse-down events consecutively,
exposes the parameter in Widget Base probe for mouse event sending,
and modifies the NumberInput probe to send the proper down and up
events.